// web/src/interfaces/Iproperties.ts
// Generated by https://quicktype.io

export interface IpropertyResponse {
  results: number;
  code: number;
  status: string;
  ok: boolean;
  data: Iproperty[];
}

export interface Iproperty {
  id: number;
  uuid: string;
  ZoneId: number;
  PropertyTypeId: number;
  OwnerId: number;
  street: string;
  number: string;
  floor: string;
  dept: string;
  isFor: string;
  state: string;
  description: string;
  createdAt?: string;
  updatedAt?: string;
  Zone?: Zone;
  PropertyType?: PropertyType;
  Owner?: Owner;
}

export interface Owner {
  fullName: string;
  phone: string;
  email: string;
}

export interface PropertyType {
  description: string;
}

export interface Zone {
  name: string;
}